==> Goldstein/2. GoldsteinData Processing.ipynb <==
cd("/Users/kyleconroy/desktop/ww_paper-1")

using Pkg
Pkg.activate("/Users/kyleconroy/desktop/ww_paper-1")

Pkg.update()
Pkg.resolve()

using DrWatson
using Revise
using JLD2
using FileIO
using CSV
using DataFrames
using Turing
using DifferentialEquations
using LogExpFunctions
using Random
using ForwardDiff
using Optim
using LineSearches
using ww_paper
using Logging
using PreallocationTools

Logging.disable_logging(Logging.Warn)

## Control Parameters
n_samples = 250 #to recreate the analysis in the paper, set to 250; 10
n_chains = 4 #to recreate the anaylsis in the paper, set to 4; 1
sim = "real"
seed = 1

if sim == "real"
  dat = CSV.read(projectdir("data/Obriendata.csv"), DataFrame)
  #dat = filter(:year_day => year_day -> year_day < 423, dat)
  ## Define Priors
include(projectdir("src", string("prior_constants_eirr_closed_LAdata", "_seed", seed, ".jl")))  ##Using the same constants as used in the paper

subset_dat = dat[:, [:new_time, :log_gene_copies]]
long_dat = DataFrames.stack(subset_dat, [:log_gene_copies])
#long_dat = filter(:value => value -> value > 0, long_dat)
data_log_copies = long_dat[:, :value]
grid_size = 1.0
end 


obstimes = long_dat[:, :new_time]
obstimes = convert(Vector{Float64}, obstimes)

# pick the change times 
if maximum(obstimes) % 7 == 0
  param_change_max = maximum(obstimes) - 7
else 
  param_change_max = maximum(obstimes)
end 
param_change_times = collect(7:7.0:param_change_max)
full_time_series = collect(minimum(obstimes):grid_size:maximum(obstimes))
outs_tmp = dualcache(zeros(6,length(full_time_series)), 10)

index = zeros(length(obstimes))
for i in 1:length(index)
    time = obstimes[i]
    index[i] = indexin(round(Int64,time), full_time_series)[1]
end 

## Define closed form solution
include(projectdir("src/newnew_closed_soln_eirr_withincid.jl"))

## Load Model
include(projectdir("src/new_bayes_eirrc_closed.jl"))

my_model = new_bayes_eirrc_closed!(
    outs_tmp, 
    data_log_copies,
    obstimes, 
    param_change_times,
    grid_size,
    index)

function optimize_many_MAP2(model, n_reps = 100, top_n = 1, verbose = true)
    lp_res = fill(-Inf, n_reps)  # Initialize lp_res as a vector with n_reps elements, each set to -Inf
    for i in eachindex(lp_res)
        if verbose
            println(i)
        end
        Random.seed!(i)
        try
            lp_res[i] = optimize(model, MAP(), LBFGS(linesearch = LineSearches.BackTracking())).lp
        catch
        end
    end
    return lp_res
end

Random.seed!(seed)

MAP_init = optimize_many_MAP(my_model, 10, 1, true)[1]

Random.seed!(seed)
MAP_noise = vcat(randn(length(MAP_init) - 1, n_chains), transpose(zeros(n_chains)))
MAP_noise = [MAP_noise[:,i] for i in 1:size(MAP_noise,2)]

init = repeat([MAP_init], n_chains) .+ 0.05 * MAP_noise

Random.seed!(seed)
posterior_samples = sample(my_model, NUTS(), MCMCThreads(), n_samples, n_chains, discard_initial = n_samples, init_params = init)

 missing_log_copies = repeat([missing], length(data_log_copies))
  
  my_model_forecast_missing = new_bayes_eirrc_closed!(
    outs_tmp, 
    missing_log_copies,
    obstimes,
    param_change_times,
    grid_size,
    index)
  
# sometimes there are NAs in the final posterior, this removes those samples
indices_to_keep = .!isnothing.(generated_quantities(my_model, posterior_samples));

posterior_samples_randn = ChainsCustomIndex(posterior_samples, indices_to_keep);

# posterior predictive output
Random.seed!(seed)
predictive_randn = predict(my_model_forecast_missing, posterior_samples_randn)

# properly scaled posterior output
Random.seed!(seed)
gq_randn = get_gq_chains(my_model, posterior_samples_randn);

CSV.write(projectdir("vignettes", string("generated_quantities", "_scenario", sim, "_seed", seed, ".csv")), DataFrame(gq_randn))
CSV.write(projectdir("vignettes", string("posterior_predictive", "_scenario", sim, "_seed", seed,  ".csv")), DataFrame(predictive_randn))

# this dataframe is saved in order to later visualize traces of the log likelihood 
posterior_df = DataFrame(posterior_samples)
CSV.write(projectdir("vignettes", string("posterior_df", "_scenario", sim, "_seed", seed, ".csv")), DataFrame(posterior_samples))

Random.seed!(seed)
prior_samples = sample(my_model, Prior(), MCMCThreads(), 400, n_chains)
    
indices_to_keep = .!isnothing.(generated_quantities(my_model, prior_samples))
prior_samples_randn = ChainsCustomIndex(prior_samples, indices_to_keep)
    
Random.seed!(seed)
prior_predictive_randn = predict(my_model_forecast_missing, prior_samples_randn)
    
Random.seed!(seed)
prior_gq_randn = get_gq_chains(my_model, prior_samples_randn)
